Add files via upload
This report dives into the inner workings of the E-Commerce code. We'll explore the challenges it addresses (problem statement), why it was created (motivation), and what it aims to achieve (objectives). We'll then unveil the step-by-step process of building it (methodology) and the tools used (hardware and software).

Next, we'll examine the data it works with (dataset description) and analyze the various outputs it generates under normal and unexpected circumstances (result analysis on various outputs and edge cases). Finally, we'll delve into the specific data structures employed in the code and the valuable lessons learned from their implementation (implementation of data structures and its learning outcomes).